### 2021-12-18 #4011 @whedon
[PRE REVIEW]: CGC: a scalable Python package for co- and tri-clustering of geo-data cubes
**Submitting author:** @fnattino (<a href="http://orcid.org/0000-0003-3286-0139">Francesco Nattino</a>)
**Repository:** <a href="https://github.com/phenology/cgc" target ="_blank">https://github.com/phenology/cgc</a>
**Version:** v0.6.1
