### 2022-01-03 #4032 @whedon
[REVIEW]: CGC: a scalable Python package for co- and tri-clustering of geo-data cubes
**Submitting author:** @fnattino (<a href="http://orcid.org/0000-0003-3286-0139">Francesco Nattino</a>)
**Repository:** https://github.com/phenology/cgc
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.6.2
