### 2022-01-13 #4032 @Subho07
The manuscript is well written and well described. Some of the required clarifications are:
1. The computation time is not compared with the existing technique. Please provide a comparison result.
2. How is the performance gets effected by the volume of data? Please provide such details in the manuscript.
