### 2022-01-13 #4032 @Narayana-Rao
The package is clean and has a well-defined purpose. The paper is also well written and tutorials are elaborative.  Please provide clarifications on the following queries.

1 ) The ``Related Software`` section is too short and requires some critical comparison and sufficient motivation.
2) The test results show a failed test associated with memory estimation (see the following screenshot), could you please fix this.
![image](https://user-images.githubusercontent.com/50410669/149345736-a20bfeee-afa5-4752-9bf5-29e734bd3e4b.png)
3) If possible add some sample data in the repository itself for tutorials and testing. Though the binder tutorials can serve the purpose, sample data may be efficient for standalone testing.
4) What is the default nature of memory and thread allocation? (I can see, we can explicitly define the parameters).
5) As also pointed out by the reviewer (@Subho07), please add some quantitative comparison on computational time and resources with existing tools/libraries.
6) I suggest adding a few lines in the manuscript about the limitations or possible futurescope of the package.
