### 2022-01-17 5d618f4
Update tests of memory estimation.
1. Updated the expected memory volumn since some cluster matrix are made boolean.
2. Update the test on the volumn difference between estimated and expected. The diff reshould be absolute value.
