### 2022-01-18 phenology/cgc#76 @fnattino
Failing test: `test_mem_estimate_coclustering_numpy`
@Narayana-Rao, thanks again for [your feedback](https://github.com/openjournals/joss-reviews/issues/4032#issuecomment-1012194082) on the package.

Concerning your point (2): we have not been able to reproduce the test failure that you observe, and while we have [fixed an issue](https://github.com/phenology/cgc/pull/75) with the same unit test, this seems to be unrelated to the error that you report. We have tried different OSs and Python versions, which we also test in our [continuous integration workflow](https://github.com/phenology/cgc/blob/master/.github/workflows/build.yml), were all tests are passing. Could you please provide some details about your setup, e.g. OS/python version/environment? Thank you so much in advance!
