+++ manuscript/paper.bib
@@ -1,4 +1,21 @@
 
+@inproceedings{Ailem:2015,
+	author = {Ailem, Melissa and Role, Fran\c{c}ois and Nadif, Mohamed},
+	title = {Co-Clustering Document-Term Matrices by Direct Maximization of Graph Modularity},
+	year = {2015},
+	isbn = {9781450337946},
+	publisher = {Association for Computing Machinery},
+	address = {New York, NY, USA},
+	url = {https://doi.org/10.1145/2806416.2806639},
+	doi = {10.1145/2806416.2806639},
+	abstract = {We present Coclus, a novel diagonal co-clustering algorithm which is able to effectively co-cluster binary or contingency matrices by directly maximizing an adapted version of the modularity measure traditionally used for networks. While some effective co-clustering algorithms already exist that use network-related measures (normalized cut, modularity), they do so by using spectral relaxations of the discrete optimization problems. In contrast, Coclus allows to get even better co-clusters by directly maximizing modularity using an iterative alternating optimization procedure. Extensive comparative experiments performed on various document-term datasets demonstrate that our algorithm is very effective, stable and outperforms other co-clustering algorithms.},
+	booktitle = {Proceedings of the 24th ACM International on Conference on Information and Knowledge Management},
+	pages = {1807–1810},
+	numpages = {4},
+	keywords = {modularity, co-clustering},
+	location = {Melbourne, Australia},
+	series = {CIKM '15}}
+
 @inproceedings{Andrienko:2009,
 	Address = {Atlantic City, NJ, USA},
 	Author = {Andrienko, Gennady and Andrienko, Natalia and Rinzivillo, Salvatore and Nanni, Mirco and Pedreschi, Dino and Giannotti, Fosca},
@@ -24,6 +41,21 @@
   	number  = {67},
   	pages   = {1919-1986},
   	url     = {http://jmlr.org/papers/v8/banerjee07a.html}}
+
+@article{Barkow:2006,
+    author = {Barkow, Simon and Bleuler, Stefan and Prelić, Amela and Zimmermann, Philip and Zitzler, Eckart},
+    title = "{BicAT: a biclustering analysis toolbox}",
+    journal = {Bioinformatics},
+    volume = {22},
+    number = {10},
+    pages = {1282-1283},
+    year = {2006},
+    month = {03},
+    abstract = "{Summary: Besides classical clustering methods such as hierarchical clustering, in recent years biclustering has become a popular approach to analyze biological data sets, e.g. gene expression data. The Biclustering Analysis Toolbox (BicAT) is a software platform for clustering-based data analysis that integrates various biclustering and clustering techniques in terms of a common graphical user interface. Furthermore, BicAT provides different facilities for data preparation, inspection and postprocessing such as discretization, filtering of biclusters according to specific criteria or gene pair analysis for constructing gene interconnection graphs. The possibility to use different biclustering algorithms inside a single graphical tool allows the user to compare clustering results and choose the algorithm that best fits a specific biological scenario. The toolbox is described in the context of gene expression analysis, but is also applicable to other types of data, e.g. data from proteomics or synthetic lethal experiments.Availability: The BicAT toolbox is freely available at  and runs on all operating systems. The Java source code of the program and a developer's guide is provided on the website as well. Therefore, users may modify the program and add further algorithms or extensions.Contact:barkow@tik.ee.ethz.ch}",
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btl099},
+    url = {https://doi.org/10.1093/bioinformatics/btl099},
[…]
