### 2022-02-07 78e91f1
add performance comparison description to docs
