### 2022-02-07 2b6668f
add defaults on implementation and number of threads employed
