### 2022-02-07 6725a1a
update how to run tests
