+++ setup.cfg
@@ -1,10 +1,6 @@
 [metadata]
 description-file = README.rst
 
-[aliases]
-# Define `python setup.py test`
-test=pytest
-
 [coverage:run]
 branch = True
 source = cgc
