+++ .github/workflows/build.yml
@@ -24,14 +24,8 @@ jobs:
           python --version
       - name: Install dependencies
         run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-      - name: Build
-        shell: bash -l {0}
-        run: |
-          python setup.py build
+          python -m pip install --upgrade pip setuptools
+          python -m pip install .[dev]
       - name: Test
-        shell: bash -l {0}
         run: |
-          pip install pytest pytest-cov pycodestyle
-          pytest --cov --cov-report term --cov-report xml --junitxml=xunit-result.xml
+          pytest -v
