### 2022-02-07 aab6039
add link to performance comparison sections
