+++ manuscript/paper.md
@@ -50,7 +50,7 @@ The exponential growth of multidimensional data referring to geographical featur
 
 In light of the eminent employability of co- and tri-clustering approaches to geospatial disciplines like geo-information science and Earth observation, and the transferability to other (geo)scientific domains, this paper presents the Clustering Geo-data Cubes (CGC) package. 
 
-As outlined below, the package aims to meet the needs of geospatial data scientists. As an example applications, we are working on the analysis of spring onset datasets at high spatial resolution and continental scale, a preview of which is presented in the CGC tutorial. While the algorithms implemented in the CGC package are limited to positive data matrices and designed to best tackle datasets with a hidden "checkerboard" structure [@Kluger:2003], they remain widely applicable and can easily be applied in other domains.
+As outlined below, the package aims to meet the needs of geospatial data scientists. An example of its application is the ongoing work on the analysis of spring onset datasets at high spatial resolution and continental scale, a preview of which is presented in the CGC tutorial. Nevertheless, the algorithms implemented in the CGC package - performing partitional clustering of positive data matrices - remain widely applicable and can easily be applied in other domains.
 
 
 # Statement of need 
@@ -89,35 +89,35 @@ For tri-clustering CGC implements the so-called Bregman cube average tri-cluster
 
 ## Cluster refinement
 
-The CGC package implements an optional, secondary cluster refinement step based on the k-means method [@Wu:2016] and optimized using the Silhouette metric [@Rousseeuw:1987] as implemented in the scikit-learn package [@Pedregosa:2011]. This secondary grouping is based on statistical properties of the co- or tri-clusters (see the [package documentation](https://cgc.readthedocs.io)) and helps to better capture patterns in the data.   
+The CGC package implements an optional, secondary cluster refinement step based on the k-means method [@Wu:2016] and optimized using the Silhouette metric [@Rousseeuw:1987] as implemented in the scikit-learn package [@Pedregosa:2011]. This secondary grouping is based on statistical properties of the co- or tri-clusters (see the [package documentation](https://cgc.readthedocs.io)) and helps to better capture patterns in the data by combining clusters and going beyond strict checkerboard structures.  
 
 
 # Related Software
 
 A number of co-/bi-clustering implementations based on a range of algorithms exist. While some of the available packages
 focus on specific applications, like gene expression data [@Barkow:2006;@Eren:2012], generic co-clustering tools include biclust 
-(R) and CoClust (Python) [@Role:2019]. The scikit-learn Python library [@Pedregosa:2011] also offers two co-clustering 
-implementations. However, most of the available algorithms target tabular data with a hidden blocked-diagonal structure, 
-where each row and column of the input matrix is assigned to only one co-cluster. In contrast, CGC's algorithm aims at 
-discovering a checkerboard-like pattern in the input data matrix, thus assigning each row and column of the 
-input data to all column and row clusters, respectively. Partitioning algorithms of this kind are suitable to 
-applications where a subset taken along one dimension is associated to multiple subsets taken along the other dimension. 
-This is often the case in spatio-temporal data, where a set of spatial elements exhibit the same behaviour within 
-multiple time windows. 
+(R) and CoClust (Python) [@Role:2019], as well as two implementations from the scikit-learn Python library [@Pedregosa:2011]. 
+However, most of the available algorithms target tabular data with a hidden blocked-diagonal structure, 
+where each row and column of the input matrix is assigned to only one co-cluster. In contrast, in spatio-temporal data a
+set of spatial elements often exhibits the same behaviour within multiple time windows and vice versa. This type of data 
+requires algorithms where a subset taken along one dimension can be associated to multiple subsets taken along the other
[…]
