### 2022-02-14 d7bdf84
Update paper.md
Summary:

- Introduction: I think the introduction was lacking what CGC is; therefore, I added two sentenced at the end (based on the summary) and linked them to the example application sentence.

- Added Numpy reference (https://numpy.org/citing-numpy/)

- Added links to pip and Zenodo pages.

- Added "function" to user-defined convergence threshold. If I recall correct it is not only a numeric value, but can also be a method, right? If not, please remove "function".

- Added secondary refinement as a unique feature.

- Added Numba reference (https://dl.acm.org/doi/10.1145/283
[…]
