### 2022-02-14 ce3c06d
Update paper.bib
- Added Harris, 2020 (Numpy)
- Added Lam, 2015 (Numba)
