+++ manuscript/paper.bib
@@ -172,6 +172,20 @@
 	Bdsk-Url-1 = {http://www.tandfonline.com/doi/abs/10.1080/13658816.2013.788249},
 	Bdsk-Url-2 = {https://doi.org/10.1080/13658816.2013.788249}}
 
+@article{Harris:2020,
+   title={Array programming with NumPy},
+   volume={585},
+   ISSN={1476-4687},
+   url={http://dx.doi.org/10.1038/s41586-020-2649-2},
+   DOI={10.1038/s41586-020-2649-2},
+   number={7825},
+   journal={Nature},
+   publisher={Springer Science and Business Media LLC},
+   author={Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del Río, Jaime Fernández and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
+   year={2020},
+   month={Sep},
+   pages={357–362}}
+
 @article{Hartigan:1972,
 	Author = {Hartigan, J. A.},
 	Doi = {10.1080/01621459.1972.10481214},
@@ -246,6 +260,23 @@
 	pages = {87--90},
 	url = {https://eprints.soton.ac.uk/403913/},
 	abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.}}
+
+@inproceedings{Lam:2015,
+	author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
+	title = {Numba: A LLVM-Based Python JIT Compiler},
+	year = {2015},
+	isbn = {9781450340052},
+	publisher = {Association for Computing Machinery},
+	address = {New York, NY, USA},
+	url = {https://doi.org/10.1145/2833157.2833162},
+	doi = {10.1145/2833157.2833162},
+	abstract = {Dynamic, interpreted languages, like Python, are attractive for domain-experts and scientists experimenting with new ideas. However, the performance of the interpreter is often a barrier when scaling to larger data sets. This paper presents a just-in-time compiler for Python that focuses in scientific and array-oriented computing. Starting with the simple syntax of Python, Numba compiles a subset of the language into efficient machine code that is comparable in performance to a traditional compiled language. In addition, we share our experience in building a JIT compiler using LLVM[1].},
+	booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+	articleno = {7},
+	numpages = {6},
+	keywords = {compiler, LLVM, Python},
+	location = {Austin, Texas},
+	series = {LLVM '15}}
 
  @misc{Merugu:2004,
 	year = {2004},
