+++ manuscript/paper.bib
@@ -261,6 +261,21 @@
 	url = {https://eprints.soton.ac.uk/403913/},
 	abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.}}
 
+@article{Kullback:1951,
+	title = {On {Information} and {Sufficiency}},
+	volume = {22},
+	issn = {0003-4851},
+	url = {http://projecteuclid.org/euclid.aoms/1177729694},
+	doi = {10.1214/aoms/1177729694},
+	language = {en},
+	number = {1},
+	urldate = {2022-02-15},
+	journal = {The Annals of Mathematical Statistics},
+	author = {Kullback, S. and Leibler, R. A.},
+	month = mar,
+	year = {1951},
+	pages = {79--86}}
+
 @inproceedings{Lam:2015,
 	author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
 	title = {Numba: A LLVM-Based Python JIT Compiler},
