### 2022-02-15 1114c34
Update paper.md
- add reference to I-divergence
- remove "function" for convergence threshold
