### 2022-02-15 9a8d8bb
Update coclustering.rst
