### 2022-02-22 #4032 @hugoledoux
> Thank you for the updates and for constructive feedback on my comments. Several things have been clarified and I think the changes have improved the overall impression greatly. I have no other comments on the package and the manuscript.

wait, could you confirm whether the changes (https://github.com/phenology/cgc/issues/76) have fixed the issue you had with a large dataset?
