### 2022-02-22 #4032 @hugoledoux
@Subho07 when you have time, could you let us know if your are satisfied with the updates made? https://github.com/openjournals/joss-reviews/issues/4032#issuecomment-1040468081
