### 2022-02-24 phenology/cgc#76 @Narayana-Rao
@fnattino, I have rerun the tests on Mac, Ubuntu and Windows 11. I didn't see any issue with mac and ubuntu. However, the same error is repeated in windows OS. Please check the attached log.txt for the details.
[log.txt](https://github.com/phenology/cgc/files/8136550/log.txt)
.
