### 2022-02-25 #4032 @fnattino
@Narayana-Rao thanks a lot for the feedback, the issue has been now solved by [phenology/cgc#82](https://github.com/phenology/cgc/pull/82)!

@hugoledoux I think all the reviewers' points have now been addressed.
