### 2022-03-09 #4032 @hugoledoux
@fnattino your paper is indeed accepted by the 2 reviewers and I am happy to start its official acceptance.

First, I fixed a few things, please accept the PR: https://github.com/phenology/cgc/pull/83
I must say that the paper is very-well written and clear, and the code and package was near perfection when you submitted. Well done.

At this point could you:
- [ ] Ensure that the copyright in the readme of your repo is 2022?
