+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,7 @@
-on: push
+on: 
+  push:
+    paths:
+      - 'manuscript/**'
 
 jobs:
   paper:
