### 2022-03-09 #4032 @fnattino
Thank you so much, very happy to hear!

> First, I fixed a few things, please accept the PR: [phenology/cgc#83](https://github.com/phenology/cgc/pull/83)

Merged!

> At this point could you:
>
> - [x]  Ensure that the copyright in the readme of your repo is 2022?
> - [x]  Make a tagged release of your software, and list the version tag of the archived version here.
> - [x]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> - [x]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> - [x]  Please list the DOI of the archived version here.

The copyright year on the README file has been fixed and I have added the ORCIDs to the Zenodo metadata, thanks for the tips!

I have made a tagged release of the software, with tag [v0.6.2](https://github.com/phenology/cgc/releases/tag/v0.6.2).
This tagged version has been archived [on Zenodo](https://zenodo.org/record/6342263) with DOI: https://doi.org/10.5281/zenodo.6342263 .

Let me know if there is anything missing on our side!
