### 2022-03-10 928f3c3
fix title capitalisation to match Zenodo
