### 2022-04-01 #4032 @fnattino
Hi @hugoledoux  just wanted to check whether everything is fine with the submission since I haven't heard from JOSS. I just noticed there were some issues with the bibliography in preparing the acceptance - but they should now be fixed.
