### 2022-04-05 #4032 @danielskatz
@hugoledoux - errors like this need to be made known to @openjournals/dev (which includes @arfon but not me)
