### 2022-04-05 #4032 @danielskatz
But just clicking through the error, I see the following issues
```
!! Invalid Crossref XML v5.3.1 in tmp/e160c54d819af470b6fd3f184cce07eb/manuscript/paper.crossref. Errors: 6.
[64](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:64)
  - 273:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet maxLength] The value has a length of 20; this exceeds the allowed maximum length of 9.
[65](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:65)
  - 273:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet pattern] The value 0162-1459, 1537-274X is not accepted by the pattern d{4}-?d{3}[dX].
[66](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:66)
  - 452:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet maxLength] The value has a length of 20; this exceeds the allowed maximum length of 9.
[67](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:67)
  - 452:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet pattern] The value 1365-8816, 1362-3087 is not accepted by the pattern d{4}-?d{3}[dX].
[68](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:68)
  - 488:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet maxLength] The value has a length of 20; this exceeds the allowed maximum length of 9.
[69](https://github.com/openjournals/joss-papers/runs/5492392521?check_suite_focus=true#step:6:69)
  - 488:0: ERROR: Element {[http://www.crossref.org/schema/5.3.1}issn:](http://www.crossref.org/schema/5.3.1%7Dissn:) [facet pattern] The value 2469-4452, 2469-4460 is not accepted by the pattern d{4}-?d{3}[dX].
```

which seems to indicate problems in the bib file, perhaps with multiple values being put in a single content-value pair.
