### 2022-04-05 #4032 @hugoledoux
@fnattino I see that you removed ISSN 4 days ago, but you forgot one: https://github.com/phenology/cgc/pull/87

If you accept it then we can try again here and it should work (touch wood).
